fix: drawers

+++ src/modules/appointment_notes/server/procedures.ts
@@ -1,6 +1,7 @@
 import prisma from "@/lib/prisma";
 import { appointmentNoteSchema } from "@/lib/schemas";
 import { createTRPCRouter, protectedProcedure } from "@/trpc/init";
+import { TRPCError } from "@trpc/server";
 import { z } from "zod";
 
 export const appointmentNotesRouter = createTRPCRouter({
@@ -38,16 +39,25 @@ export const appointmentNotesRouter = createTRPCRouter({
 			return c;
 		}),
 	delete: protectedProcedure
-		.input(z.object({ noteId: z.string().cuid2() }))
+		.input(z.object({ id: z.string().cuid2().nullable() }))
 		.mutation(async ({ input }) => {
-			const { noteId } = input;
+			const { id } = input;
 
-			const c = await prisma.appointmentNote.delete({
+			if (!id) {
+				throw new TRPCError({ code: "NOT_FOUND", message: "Missing id" });
+			}
+
+			const note = await prisma.appointmentNote.delete({
 				where: {
-					id: noteId,
+					id,
 				},
 			});
-			return c;
+
+			if (!note) {
+				throw new TRPCError({ code: "NOT_FOUND" });
+			}
+
+			return note;
 		}),
 	getNotesByAppointmentId: protectedProcedure
 		.input(z.object({ appointmentId: z.string().cuid2() }))
@@ -61,5 +71,26 @@ export const appointmentNotesRouter = createTRPCRouter({
 			});
 
 			return notes;
+		}),
+	getById: protectedProcedure
+		.input(z.object({ id: z.string().cuid2().nullable() }))
+		.query(async ({ input }) => {
+			const { id } = input;
+
+			if (!id) {
+				throw new TRPCError({ code: "NOT_FOUND", message: "Missing id" });
+			}
+
+			const note = await prisma.appointmentNote.findFirst({
+				where: {
+					id,
+				},
+			});
+
+			if (!note) {
+				throw new TRPCError({ code: "NOT_FOUND" });
+			}
+
+			return note;
 		}),
 });
